use line notify

+++ lib/useReservations.ts
@@ -1,3 +1,5 @@
+import { lineNotify } from "./linenotify";
+
 export interface Reservation {
   id: string;
   userName: string;
@@ -19,32 +21,38 @@ export const getReservations = async (microcmsClient, filters: string) => {
   return res.contents;
 }
 
-export const createReservation = (microcmsClient, reservation: Reservation) => {
+export const createReservation = (microcmsClient, reservation: Reservation, success: () => {}) => {
+  if(!confirm("Do you reserve time?"))return;
   microcmsClient.create({
     endpoint: 'reservations',
     content: reservation,
   })
+  .then(success)
   .catch((err) => console.error(err));
 
   return reservation;
 }
 
-export const deleteReservation = (microcmsClient, reservation: Reservation) => {
+export const deleteReservation = (microcmsClient, reservation: Reservation, success: () => {}) => {
+  if(!confirm("Do you delete this reservation?"))return;
   microcmsClient.delete({
     endpoint: 'reservations',
     contentId: reservation.id,
   })
+  .then(success)
   .catch((err) => console.error(err));
 
   return reservation;
 }
 
 export const updateReservation = (microcmsClient, reservation: Reservation) => {
-  microcmsClient.create({
-    endpoint: 'reservations',
-    content: reservation,
-  })
-  .catch((err) => console.error(err));
+  microcmsClient
+    .update({
+      endpoint: `reservations/${reservation.id}`,
+      content: reservation,
+    })
+    .then((res) => console.log(res.id))
+    .catch((err) => console.error(err));
 
   return reservation;
 }
